requirements.txt
apache-airflow==2.9.2
cassandra_driver==3.29.1
findspark==2.0.1
flasgger==0.9.7.1
Flask==2.2.5
kafka_python==2.0.2
pyspark==3.5.1
Requests==2.32.3
